+++ app/scripts/background.js
@@ -76,12 +76,19 @@ var providerOpts = {
 var provider = MetaMaskProvider(providerOpts)
 var web3 = new Web3(provider)
 idStore.web3 = web3
-idStore.getNetwork(3)
+idStore.getNetwork()
 
 // log new blocks
 provider.on('block', function(block){
   console.log('BLOCK CHANGED:', '#'+block.number.toString('hex'), '0x'+block.hash.toString('hex'))
+
+  // Check network when restoring connectivity:
+  if (idStore._currentState.network === 'loading') {
+    idStore.getNetwork()
+  }
 })
+
+provider.on('error', idStore.getNetwork.bind(idStore))
 
 var ethStore = new EthStore(provider)
 idStore.setStore(ethStore)
@@ -145,7 +152,7 @@ function setupPublicConfig(stream){
 }
 
 function setupProviderConnection(stream, originDomain){
-  
+
   stream.on('data', function onRpcRequest(payload){
     // Append origin to rpc payload
     payload.origin = originDomain
@@ -195,6 +202,8 @@ function setupControllerConnection(stream){
     exportAccount:      idStore.exportAccount.bind(idStore),
     revealAccount:      idStore.revealAccount.bind(idStore),
     saveAccountLabel:   idStore.saveAccountLabel.bind(idStore),
+    tryPassword:        idStore.tryPassword.bind(idStore),
+    recoverSeed:        idStore.recoverSeed.bind(idStore),
   })
   stream.pipe(dnode).pipe(stream)
   dnode.on('remote', function(remote){
@@ -246,7 +255,7 @@ function newUnsignedTransaction(txParams, cb){
     })
     var txId = idStore.addUnconfirmedTransaction(txParams, cb)
   } else {
-    addUnconfirmedTx(txParams, cb)    
+    addUnconfirmedTx(txParams, cb)
   }
 }
 
@@ -258,7 +267,7 @@ function newUnsignedMessage(msgParams, cb){
     })
     var msgId = idStore.addUnconfirmedMessage(msgParams, cb)
   } else {
-    addUnconfirmedMsg(msgParams, cb)    
+    addUnconfirmedMsg(msgParams, cb)
   }
 }
 
@@ -290,13 +299,13 @@ function addUnconfirmedMsg(msgParams, cb){
 function setRpcTarget(rpcTarget){
   configManager.setRpcTarget(rpcTarget)
   chrome.runtime.reload()
-  idStore.getNetwork(3) // 3 retry attempts
+  idStore.getNetwork()
 }
 
 function setProviderType(type) {
   configManager.setProviderType(type)
   chrome.runtime.reload()
-  idStore.getNetwork(3)
+  idStore.getNetwork()
 }
 
 function useEtherscanProvider() {